refactor: compare changes with base-parachain
