+++ Cargo.toml
@@ -21,9 +21,10 @@ parity-scale-codec = { version = "3.6.12", default-features = false, features = 
 	"max-encoded-len",
 ] }
 color-print = "0.3.4"
+docify = "0.2.8"
 hex-literal = "0.4.1"
-log = { version = "0.4.20", default-features = false }
-scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
+log = { version = "0.4.21", default-features = false }
+scale-info = { version = "2.11.1", default-features = false, features = ["derive"] }
 smallvec = "1.11.2"
 serde = "1.0.197"
 clap = { version = "4.5.3", features = ["derive"] }
