+++ node/src/cli.rs
@@ -40,10 +40,6 @@ pub enum Subcommand {
     #[command(subcommand)]
     Benchmark(frame_benchmarking_cli::BenchmarkCmd),
 
-    /// Try-runtime has migrated to a standalone
-    /// [CLI](<https://github.com/paritytech/try-runtime-cli>). The subcommand exists as a stub and
-    /// deprecation notice. It will be removed entirely some time after January 2024.
-    TryRuntime,
     /// Db meta columns information.
     FrontierDb(fc_cli::FrontierDbCmd),
 }
